refactor train.py
